requirements.txt
aiofiles==23.2.1
altair==5.5.0
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.8.0
apispec==6.8.2
attrs==25.3.0
backports-datetime-fromisoformat==2.0.3
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
faiss-cpu==1.7.4
fastapi==0.115.8
ffmpy==0.5.0
filelock==3.17.0
Flask==2.3.3
flask-apispec==0.11.4
Flask-Cors==3.0.10
Flask-RESTful==0.3.10
Flask-SocketIO==4.3.2
fsspec==2025.2.0
gevent==22.10.2
gevent-websocket==0.10.1
gitdb==4.0.12
GitPython==3.1.44
gradio==5.14.0
gradio_client==1.7.0
greenlet==3.2.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.17.3
idna==3.10
importlib-metadata==6.11.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==4.0.0
marshmallow-enum==1.5.1
marshmallow-jsonapi==0.24.0
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
narwhals==1.41.1
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==1.24.3
openai-whisper==20240930
openfabric-pysdk==0.2.7
orjson==3.10.15
packaging==23.2
pandas==2.2.3
pickleDB==0.9.2
Pillow==10.0.0
protobuf==4.25.8
pyarrow==19.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==3.14.2
python-multipart==0.0.20
python-socketio==4.6.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==25.1.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.25.1
ruff==0.9.4
runstats==2.0.0
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.2.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socketIO-client==0.7.2
starlette==0.45.3
streamlit==1.28.0
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.14.1
toml==0.10.2
tomlkit==0.13.2
torch==2.0.1
torchvision==0.15.2
tornado==6.5.1
tqdm==4.67.1
transformers==4.35.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
validators==0.35.0
watchdog==6.0.0
webargs==8.7.0
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.22.0
zope.event==5.0
zope.interface==7.2
